move everything into the -rand crate
