+++ forrustts-genetic-maps-rand/tests/tests.rs
@@ -1,4 +1,4 @@
-use forrustts_genetic_maps::GeneticMapBuilder;
+use forrustts_genetic_maps_rand::GeneticMapBuilder;
 use forrustts_genetic_maps_rand::GeneticMap;
 
 #[test]
